+++ tests/unit/test_td/_context.py
@@ -8,8 +8,12 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__),
                                                 '../../..')))
 
 # pylint: disable=wrong-import-position, unused-import, wrong-import-order
-from wtalib._td._array import MaskedArray  # noqa: E402, F401
+from wtalib._td._array import (  # noqa: E402, F401
+    LogicalBinaryOperator,
+    LogicalUnaryOperator,
+    MaskedArray,
+)
 from wtalib._td._index import TimeIndex  # noqa: E402, F401
-from wtalib._td._series import TimeSeries  # noqa: E402, F401
+from wtalib._td._series import BooleanTimeSeries, TimeSeries  # noqa: E402, F401
 
 # pylint: enable=wrong-import-position, unused-import, wrong-import-order