+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **TimeIndex**, an immutalbe sequence of datetimes, used as index of time-indexing data.
 - **MaskedArray**, an array with masks indicating which elements ara N/A.
 - **TimeSeries**, a sequence of data points indexed by time.
+- **BooleanTimeSeries**, a time-series of boolean data.